functions/session-mapper/requirements.txt
flask==3.1.1
functions-framework==3.8.3
requests==2.32.3
google-auth==2.39.0